Add is_stack field to Device db model
Adds possiblity to define a device as a stack through a
is_stack field.
Checking if stackmemebers exists is not good enough
to decide if a device is a stack. The stackmember model
is used to map a serialnumber to a position ID. This is
not required for the stack to work, so a stack device does not
need any stackmembers to be defined in cnaas_nms to still work.
Because of this an explicit way to say if something is a stack or not is
useful. Also added the field to the validation method and flask model