Make template secrets available as dict in context
Because:
- Sometimes, we need to construct secret names from variables that
  are available in the template context. This isn't possible when
  accessing top-level/global context variables, but it is possible
  to construct dictionary keys from variables.

Example:

Looping through a list of tacacs+ servers, each of them have a distinct
secret key, which we need to fetch from the environment. I.e. the
environment variables will be different for each server. For the
tacacs+ server 10.0.1.1 we want to fetch the env var
TEMPLATE_SECRET_TACPLUS_KEY_10_0_1_1. Jinja doesn't allow us to
construct this name, but we can get it using something like this:
{{ TEMPLATE_SECRET['TACPLUS_KEY_' ~ ipaddr] }}