Require unique hardware_id only within a device
Previously hardware_id had to be unique among all stackmembers,
regardless of which stack they belonged to. As you might want to
configure a stack with stackmembers before actually adding them to the stack,
it is possible for duplicate hardware_id to exist between stacks.

The line of thinking that hardware_id must be unique for all Stackmembers
would in my mind also require it to be unique amongs all Devices, as these
are also switches.