Remove old teardown and unused setup instance
There was some code in setup that started some db instance, but
looking at the code for it it didnt actually do anything, and still requires
a db to be running locally.

In teardown, device in the db is reset i assume to make sure
it is in the same state for all the tests.
this is removed, as later commits will make it so no pre-existing
data is expected to exist in the db.

the teset_5 function also had identical code to what was in
teardown, without any sort of asserts following it. I assume
this is also some sort of teardown that was not supposed to be there,
so it is removed as well