Improve error message for import error
If jinja_filters is actually found, but its sub-dependencies cannot be
imported, the actual culprit would be masked. This adds the exception
string to the output.

Our users were told jinja_filters was not found, when the actual
problem was that they had not installed the cachetools module, which
our current filter set depends on.