Make ipv4_to_ipv6 output "human-readable" addrs
The requirements keep changing back and forth, but apparently, it
was still important that the resulting IPv6 address was somewhat
human-readably similar to the original IPv4 address.

Also, it isn't always necessary to keep the full IPv4 address inside
the IPv6 address. Mostly, the host address from the IPv4 subnetwork
should be readable as the same on the corresponding IPV6 subnetwork,
so an option to keep only a number of IPv4 octets in the result was
added.