Build the docker image from the source repo
Whether the source repo for a triggered workflow is the upstream repo or
a fork, this should ensure the docker image is built from the right
version.

I would, of course, prefer a version that builds the image from the
current checkout, rather than make its own, since this would be more
compatible with typival development workflows, but that can come later.