Fix broken COVERAGE option check
The default `docker-compose.yml` sets the environment variable COVERAGE to
an empty value. However, at runtime, the value of the environment
variable is never verified: The presence of the variable in the
environment is enough to enable coverage reporting in the code base.

Due to this, there is also no way to override the COVERAGE setting in
e.g. `docker-compose.override.yml`, since you cannot get docker-compose
to *remove* the variable from the environment, and you cannot give it a
value that turns coverage off.

This appears to cause coverage to be collected even in production
settings, which is a performance no-no. Also, in our case, there is some
privilege or dependency problem that crashes the API server in
production, as the coverage module is unable to create a coverage report
SQLite database in the configured/preferred location.