Disable JWT_ENABLED in test suite
Just importing the cnaas_nms.app module runs sys.exit() (a big no no)
if the JWT certificate file cannot be found. The default path for
this certificate is a global path, whic is also a no-no, especially
during tests.

This simply disables JWT during the running of all tests under pytest.
A better solution would be to construct ApiSettings specifically for a
test run.  Unit tests should not be dependent on external files, and
any integration tests that need a certificate for testing should
generate a certificate for each test run.