Wait longer for PostgreSQL to come up
Some tests are failing when singled out for a test session.

This is because there is an apparent lag between the PostgreSQL
container responding to TCP port 5432 and actually being ready to serve
database connections. Sleeping for 1 second is the naive solution here,
but it's okay, since it only happens the first time the fixture is used
in a test session.  A more complete solution would check that we can
actually establish a PostgreSQL client